Add sticky svg
